fix(review): split the portwatch transport tests and harden two rungs
Code review of #7672 surfaced four validated findings, all test-only:

- The suite file crossed 1000 lines; move the reference-pagination
  describe with its exclusive fixtures into
  tests/portwatch-port-refs-transport.test.mjs, matching the existing
  per-concern portwatch layout (original returns to 646 lines).
- Six uniform pagination tests re-hand-rolled the transport stub; one
  paginationTransport() helper now covers them (offset-parse copies
  8 -> 4 in the moved file), while the stateful, abort, and
  default-transport cases keep hand-rolled stubs because their shape
  is the point of the test.
- The caller-abort test threw a plain Error, so deleting the
  `if (signal?.aborted) throw err` guard it names left it green (the
  error re-routed through the !isTimeoutLike rung instead). It now
  throws the realistic DOMException AbortError class the guard exists
  to intercept; mutation-verified red without the guard, green with it.
- The non-timeout-like transport-throw rung (`if (!isTimeoutLike)`)
  had no test; a permanent transport error now asserts identity
  rejection with zero proxy calls. Mutation-verified both directions.

Also pins the 429 test's proxy URL at offset 1, matching its title and
the HTTP-200 sibling.

Refs #7672
