+++ tests/portwatch-port-activity-recovery.test.mjs
@@ -1,54 +1,10 @@
 import { describe, it } from 'node:test';
 import assert from 'node:assert/strict';
-import { readFile } from 'node:fs/promises';
 
 import * as portwatchSeed from '../scripts/seed-portwatch-port-activity.mjs';
-import { CHROME_UA } from '../scripts/_seed-utils.mjs';
 
 const { orderColdFetchQueue } = portwatchSeed;
 const DAY = 86_400_000;
-// Sanitized from a real Railway deployment. Its _comment records that only
-// the HTTP 200 envelope and message were confirmed upstream — other provider
-// fields are unconfirmed rather than observed absent, so do not treat this
-// fixture as proof of the full error shape.
-const arcgisRateLimitFixture = JSON.parse(await readFile(
-  new URL('./fixtures/portwatch-arcgis-too-many-requests.json', import.meta.url),
-  'utf8',
-));
-
-function arcgisJson(body, status = 200) {
-  return new Response(JSON.stringify(body), {
-    status,
-    headers: { 'Content-Type': 'application/json' },
-  });
-}
-
-// The two reference pages every pagination case below walks: a first page
-// that reports more to come, and a final page that closes the loop. Only the
-// failure injected between them varies per test. arcgisJson serializes its
-// argument, so sharing these objects cannot leak state across tests.
-const FIRST_PAGE = {
-  features: [{ attributes: { portid: 'us-lax', ISO3: 'USA', lat: 33.7, lon: -118.2 } }],
-  exceededTransferLimit: true,
-};
-const FINAL_PAGE = {
-  features: [{ attributes: { portid: 'cy-lca', ISO3: 'CYP', lat: 34.9, lon: 33.6 } }],
-  exceededTransferLimit: false,
-};
-
-// Drive one reference page whose second offset always returns `body`, and
-// return the message fetchWithTimeout's parser threw for it.
-async function rejectedRefsError(body) {
-  const fetchFn = async (url) => {
-    const offset = Number(new URL(url).searchParams.get('resultOffset'));
-    return arcgisJson(offset === 0 ? FIRST_PAGE : body);
-  };
-  const err = await portwatchSeed
-    .fetchAllPortRefs({ fetchFn, sleepFn: async () => {} })
-    .then(() => null, (e) => e);
-  assert.ok(err, 'expected fetchAllPortRefs to reject');
-  return err.message;
-}
 
 function cachedCountry(iso2, cacheWrittenAt = 0) {
   return {
@@ -66,325 +22,6 @@ function activityResult(portAccumMap, currentWindowRowCount = portAccumMap.size)
   return { portAccumMap, currentWindowRowCount };
 }
 
-describe('PortWatch reference pagination recovery', () => {
-  it('parses and retries a rate-limited HTTP 200 page without restarting pagination', async () => {
-    assert.equal(typeof arcgisRateLimitFixture.error, 'object');
-    assert.equal(
-      arcgisRateLimitFixture.error.message,
-      'Unable to perform query. Too many requests.',
-    );
-
-    const requestedOffsets = [];
-    const requestedInits = [];
-    const proxyCalls = [];
-    const fetchFn = async (url, init) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-      requestedInits.push(init);
-
-      return arcgisJson(offset === 0 ? FIRST_PAGE : arcgisRateLimitFixture);
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({
-      fetchFn,
-      proxyRetryFn: async (...args) => {
-        proxyCalls.push(args);
-        return FINAL_PAGE;
-      },
-    });
-
-    assert.deepEqual(requestedOffsets, [0, 1]);
-    assert.deepEqual([...refsByIso3.get('USA').keys()], ['us-lax']);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-    assert.equal(proxyCalls.length, 1);
-    assert.equal(new URL(proxyCalls[0][0]).searchParams.get('resultOffset'), '1');
-    assert.equal(proxyCalls[0][1], 'HTTP 200 rate-limited');
-    assert.ok(proxyCalls[0][2].signal instanceof AbortSignal);
-
-    // The seam is only faithful if it hands the transport what production
-    // hands it. fetchWithTimeout builds these headers and the combined abort
-    // signal itself, so without this assertion the required User-Agent or the
-    // abort wiring could be dropped and every test here would stay green.
-    assert.equal(requestedInits.length, 2);
-    for (const init of requestedInits) {
-      assert.equal(init.headers['User-Agent'], CHROME_UA);
-      assert.equal(init.headers.Accept, 'application/json');
-      assert.ok(init.signal instanceof AbortSignal);
-    }
-  });
-
-  it('bounds repeated HTTP 200 rate-limit errors to one same-page retry', async () => {
-    const requestedOffsets = [];
-    const proxyCalls = [];
-    const sleepCalls = [];
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-      if (offset === 0) {
-        return arcgisJson(FIRST_PAGE);
-      }
-      return arcgisJson(arcgisRateLimitFixture);
-    };
-
-    await assert.rejects(
-      portwatchSeed.fetchAllPortRefs({
-        fetchFn,
-        proxyRetryFn: async (...args) => {
-          proxyCalls.push(args);
-          throw portwatchSeed.createArcgisProxyError(args[1], 'Too many requests');
-        },
-        sleepFn: async (...args) => {
-          sleepCalls.push(args);
-        },
-      }),
-      /ArcGIS error \(via proxy after HTTP 200 rate-limited\): Too many requests/,
-    );
-
-    assert.deepEqual(requestedOffsets, [0, 1, 1]);
-    assert.equal(proxyCalls.length, 2);
-    assert.equal(sleepCalls.length, 1);
-    assert.equal(sleepCalls[0][0], 2_000);
-  });
-
-  it('retries a code-only rate-limit envelope that carries no message', async () => {
-    // ArcGIS also emits `{"error":{"code":N}}` with no message field (the
-    // shape documented on the proxy parser). The direct parser must surface
-    // the code so the retry classifier still sees a rate limit, rather than
-    // throwing "ArcGIS error: undefined" and abandoning the page.
-    const requestedOffsets = [];
-    const proxyCalls = [];
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-
-      return arcgisJson(offset === 0 ? FIRST_PAGE : { error: { code: 429 } });
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({
-      fetchFn,
-      proxyRetryFn: async (...args) => {
-        proxyCalls.push(args);
-        return FINAL_PAGE;
-      },
-    });
-
-    assert.deepEqual(requestedOffsets, [0, 1]);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-    assert.equal(proxyCalls.length, 1);
-    assert.equal(proxyCalls[0][1], 'HTTP 200 rate-limited');
-  });
-
-  it('routes the invalid-params retry backoff through the injected sleep', async (t) => {
-    // The other ArcGIS failure class. Its 500ms backoff lives in
-    // fetchWithRetryOnInvalidParams rather than retryRateLimited, so it has
-    // to reach the same injected sleepFn -- otherwise this branch cannot be
-    // covered without paying real wall-clock on every run.
-    portwatchSeed._resetInvalidParamsErrorCount();
-    t.after(() => portwatchSeed._resetInvalidParamsErrorCount());
-
-    const requestedOffsets = [];
-    const sleepCalls = [];
-    let offsetOneAttempts = 0;
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-
-      if (offset === 0) {
-        return arcgisJson(FIRST_PAGE);
-      }
-
-      offsetOneAttempts += 1;
-      if (offsetOneAttempts === 1) {
-        return arcgisJson({
-          error: { message: 'Cannot perform query. Invalid query parameters.' },
-        });
-      }
-      return arcgisJson(FINAL_PAGE);
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({
-      fetchFn,
-      sleepFn: async (...args) => {
-        sleepCalls.push(args);
-      },
-    });
-
-    assert.deepEqual(requestedOffsets, [0, 1, 1]);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-    assert.deepEqual(sleepCalls.map(([ms]) => ms), [500]);
-  });
-
-  // The three rungs of the error ladder, pinned separately. The retry
-  // classifier reads the thrown message, and `{"code":429}` stringifies to
-  // something that still matches /\b429\b/ -- so a test that only proves a
-  // retry fired cannot tell `?? code` from `?? JSON.stringify`. These assert
-  // the message itself, on codes that do NOT trip a retry.
-  it('prefers message over code, and code over the stringified envelope', async () => {
-    const messageAndCode = await rejectedRefsError({
-      error: { code: 400, message: 'Cannot perform query. Bad field.' },
-    });
-    assert.equal(messageAndCode, 'ArcGIS error: Cannot perform query. Bad field.');
-
-    const codeOnly = await rejectedRefsError({ error: { code: 400 } });
-    assert.equal(codeOnly, 'ArcGIS error: 400');
-
-    const neither = await rejectedRefsError({ error: { details: ['nope'] } });
-    assert.equal(neither, 'ArcGIS error: {"details":["nope"]}');
-  });
-
-  // The status-429 branch, not the HTTP-200 rate-limit envelope the earlier
-  // tests drive: no test above hands the transport a non-200 status, so
-  // deleting `if (resp.status === 429)` leaves the whole suite green
-  // (verified by mutation on #7539's head).
-  it('routes an HTTP 429 status through the proxy transport at the same offset', async () => {
-    const requestedOffsets = [];
-    const proxyCalls = [];
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-      return offset === 0
-        ? arcgisJson(FIRST_PAGE)
-        : arcgisJson(arcgisRateLimitFixture, 429);
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({
-      fetchFn,
-      proxyRetryFn: async (...args) => {
-        proxyCalls.push(args);
-        return FINAL_PAGE;
-      },
-    });
-
-    assert.deepEqual(requestedOffsets, [0, 1]);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-    assert.equal(proxyCalls.length, 1);
-    assert.equal(proxyCalls[0][1], 'HTTP 429 rate-limited');
-    assert.ok(proxyCalls[0][2].signal instanceof AbortSignal);
-  });
-
-  it('rejects on a non-429 HTTP error status without retrying or proxying', async () => {
-    const requestedOffsets = [];
-    const proxyCalls = [];
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-      return offset === 0
-        ? arcgisJson(FIRST_PAGE)
-        : arcgisJson({ error: { message: 'Service unavailable.' } }, 503);
-    };
-
-    await assert.rejects(
-      portwatchSeed.fetchAllPortRefs({
-        fetchFn,
-        proxyRetryFn: async (...args) => {
-          proxyCalls.push(args);
-          return FINAL_PAGE;
-        },
-      }),
-      /ArcGIS HTTP 503 for /,
-    );
-
-    // retryRateLimited only retries the rate-limit class, and the invalid-
-    // params circuit breaker only matches its own message — a bare HTTP
-    // error status must surface as-is after one attempt per offset.
-    assert.deepEqual(requestedOffsets, [0, 1]);
-    assert.equal(proxyCalls.length, 0);
-  });
-
-  it('falls back to the proxy transport when the direct transport throws a timeout-like error', async () => {
-    // The WM 2026-05-13 silent-stall mode: ArcGIS never answers, the
-    // transport throws instead of returning a response. The thrown message
-    // must match an isTimeoutLike class for the proxy fallback to open.
-    const requestedOffsets = [];
-    const proxyCalls = [];
-    const fetchFn = async (url) => {
-      const offset = Number(new URL(url).searchParams.get('resultOffset'));
-      requestedOffsets.push(offset);
-      if (offset === 0) return arcgisJson(FIRST_PAGE);
-      throw new TypeError('fetch failed');
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({
-      fetchFn,
-      proxyRetryFn: async (...args) => {
-        proxyCalls.push(args);
-        return FINAL_PAGE;
-      },
-    });
-
-    assert.deepEqual(requestedOffsets, [0, 1]);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-    assert.equal(proxyCalls.length, 1);
-    assert.equal(proxyCalls[0][1], 'direct TypeError');
-  });
-
-  it('propagates a caller abort from the transport catch without proxy fallback', async () => {
-    // A real cancellation (SIGTERM / per-country timeout) must rethrow the
-    // caller's own reason: routing it to the proxy retry would spend a
-    // paid request on work the run already gave up on.
-    const proxyCalls = [];
-    const controller = new AbortController();
-    const abortReason = new Error('run cancelled');
-    const fetchFn = async () => {
-      controller.abort(abortReason);
-      throw controller.signal.reason;
-    };
-
-    await assert.rejects(
-      portwatchSeed.fetchAllPortRefs({
-        signal: controller.signal,
-        fetchFn,
-        proxyRetryFn: async (...args) => {
-          proxyCalls.push(args);
-          return FINAL_PAGE;
-        },
-      }),
-      (err) => err === abortReason,
-    );
-    assert.equal(proxyCalls.length, 0);
-  });
-
-  it('rejects before dialling out when the run signal is already aborted', async () => {
-    const directCalls = [];
-    const controller = new AbortController();
-    const abortReason = new Error('run cancelled');
-    controller.abort(abortReason);
-
-    await assert.rejects(
-      portwatchSeed.fetchAllPortRefs({
-        signal: controller.signal,
-        fetchFn: async (...args) => {
-          directCalls.push(args);
-          return arcgisJson(FIRST_PAGE);
-        },
-      }),
-      (err) => err === abortReason,
-    );
-    assert.equal(directCalls.length, 0);
-  });
-
-  it('wires the default direct transport to globalThis.fetch', async (t) => {
-    // defaultFetch must read globalThis.fetch at call time rather than
-    // binding the original once, so swapping the global is observable here.
-    const originalFetch = globalThis.fetch;
-    const directCalls = [];
-    t.after(() => {
-      globalThis.fetch = originalFetch;
-    });
-    globalThis.fetch = async (url, init) => {
-      directCalls.push({ url: String(url), init });
-      return arcgisJson(FINAL_PAGE);
-    };
-
-    const refsByIso3 = await portwatchSeed.fetchAllPortRefs({});
-
-    assert.equal(directCalls.length, 1);
-    assert.equal(directCalls[0].init.headers['User-Agent'], CHROME_UA);
-    assert.equal(directCalls[0].init.headers.Accept, 'application/json');
-    assert.ok(directCalls[0].init.signal instanceof AbortSignal);
-    assert.deepEqual([...refsByIso3.get('CYP').keys()], ['cy-lca']);
-  });
-});
 
 describe('PortWatch activity observation recovery', () => {
   it('keeps an explicit null max date separate from a missing aggregate row', () => {
